london_app_dev/requirements.txt
Django==5.1.6
djangorestframework==3.15.2
django-cors-headers==4.7.0
psycopg[c]==3.2.5